Fixed bug - integer overlooked during int to long conversion
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@21062 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Service/src/main/java/org/gusdb/wdk/service/request/user/UserDatasetShareRequest.java
@@ -99,7 +99,7 @@ public class UserDatasetShareRequest {
           String usersJson = null;
           try {
             usersJson = usersJsonArray.toString();
-            CollectionType setType = mapper.getTypeFactory().constructCollectionType(Set.class, Integer.class);
+            CollectionType setType = mapper.getTypeFactory().constructCollectionType(Set.class, Long.class);
             Set<Long> users = mapper.readValue(usersJson, setType);
             innerMap.put(dataset, users);  
           }
